Update main.c
I thought it was going to be more intricate and a little less obvious. Also, the fact that the question states "You must modify as few characters as possible." I find that rather hindering my creativity. Surely there are more ways to solve this problem, yet you only allow very few changes to be made. If maybe the odds of changing characters were in our favor, maybe the outcome of students' creative side will be apparent.